requirements.txt
numpy
pandas
scikit-learn
thefuzz
sentence-transformers
torch
python-dotenv
